replace existing MsgpackEncoder
