+++ tests/benchmarks/test_encoding.py
@@ -1,16 +1,14 @@
-import random
-import string
-
 import msgpack
 from msgpack.fallback import Packer
 import pytest
 
-from ddtrace import Span, Tracer
-from ddtrace.encoding import _EncoderBase, MsgpackEncoder, TraceMsgpackEncoder
+from ddtrace.encoding import _EncoderBase, MsgpackEncoder
+
+from ..test_encoders import RefMsgpackEncoder, gen_trace
 
 
-msgpack_encoder = MsgpackEncoder()
-trace_encoder = TraceMsgpackEncoder()
+msgpack_encoder = RefMsgpackEncoder()
+trace_encoder = MsgpackEncoder()
 
 
 class PPMsgpackEncoder(_EncoderBase):
@@ -25,56 +23,6 @@ class PPMsgpackEncoder(_EncoderBase):
         if msgpack.version[:2] < (0, 6):
             return msgpack.unpackb(data)
         return msgpack.unpackb(data, raw=True)
-
-
-def rands(size=6, chars=string.ascii_uppercase + string.digits):
-    return "".join(random.choice(chars) for _ in range(size))
-
-
-def gen_span(length=None, **span_attrs):
-    # Helper to generate spans
-    name = span_attrs.pop("name", None)
-    if name is None:
-        name = "a" * length
-
-    span = Span(None, **span_attrs)
-
-    for attr in span_attrs:
-        if hasattr(span, attr):
-            setattr(span, attr, attr)
-        else:
-            pass
-
-    if length is not None:
-        pass
-
-
-def gen_trace(nspans=1000, ntags=50, key_size=15, value_size=20, nmetrics=10):
-    t = Tracer()
-
-    root = None
-    trace = []
-    for i in range(0, nspans):
-        parent_id = root.span_id if root else None
-        with Span(
-            t, "span_name", resource="/fsdlajfdlaj/afdasd%s" % i, service="myservice", parent_id=parent_id,
-        ) as span:
-            span._parent = root
-            span.set_tags({rands(key_size): rands(value_size) for _ in range(0, ntags)})
-
-            # only apply a span type to the root span
-            if not root:
-                span.span_type = "web"
-
-            for _ in range(0, nmetrics):
-                span.set_tag(rands(key_size), random.randint(0, 2 ** 16))
-
-            trace.append(span)
-
-            if not root:
-                root = span
-
-    return trace
 
 
 trace_large = gen_trace(nspans=1000)
